Move tests for the moved code
